Include video button into header
