+++ plugiamo/src/app/content/scripted-chat/components/message-types/video-message.js
@@ -1,6 +1,6 @@
 import mixpanel from 'ext/mixpanel'
-import Modal from 'shared/modal'
 import styled from 'styled-components'
+import VideoModal from 'shared/video-modal'
 import { compose, withHandlers, withProps, withState } from 'recompose'
 import { h } from 'preact'
 import { IconPlayButton } from 'plugin-base'
@@ -24,11 +24,6 @@ const VideoMessage = compose(
     youtubeUrl: `https://www.youtube.com/watch?v=${youtubeId}`,
   })),
   withHandlers({
-    closeModal: ({ setIsOpen, youtubeUrl }) => event => {
-      event.stopPropagation()
-      setIsOpen(false)
-      mixpanel.track('Closed Video', { hostname: location.hostname, url: youtubeUrl })
-    },
     openModal: ({ setIsOpen, youtubeUrl }) => () => {
       setIsOpen(true)
       mixpanel.track('Open Video', { hostname: location.hostname, url: youtubeUrl })
@@ -38,27 +33,13 @@ const VideoMessage = compose(
     onClick: ({ openModal }) => openModal,
     onKeyUp: ({ openModal }) => event => (event.key === 'Enter' ? openModal() : undefined),
   })
-)(styled(({ className, closeModal, isOpen, onClick, onKeyUp, youtubeEmbedUrl, youtubePreviewImageUrl }) => (
+)(styled(({ className, setIsOpen, isOpen, onClick, onKeyUp, youtubeEmbedUrl, youtubePreviewImageUrl }) => (
   <div className={className} onClick={onClick} onKeyUp={onKeyUp} role="button" tabIndex={0}>
     <img alt="" src={youtubePreviewImageUrl} style={{ maxWidth: '100%' }} />
     <IconContainer>
       <IconPlayButton />
     </IconContainer>
-    <Modal allowBackgroundClose closeModal={closeModal} isOpen={isOpen}>
-      <div style={{ width: '100%', paddingBottom: '56.25%', position: 'relative' }}>
-        <div style={{ position: 'absolute', top: 0, bottom: 0, left: 0, right: 0 }}>
-          <iframe
-            allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture"
-            allowFullScreen="1"
-            frameBorder="0"
-            height="100%"
-            src={youtubeEmbedUrl}
-            title="Video player"
-            width="100%"
-          />
-        </div>
-      </div>
-    </Modal>
+    <VideoModal isOpen={isOpen} setIsOpen={setIsOpen} url={youtubeEmbedUrl} />
   </div>
 ))`
   cursor: pointer;
